refactor(inventory): cleanup product aggregate and features (#50)
## Description
This PR refactors `Product` aggregate, it's features and configurations.
Initial migration is recreated to fit the new version of the aggregate.

### Key changes
- Fields like `Description`, `Price`, `Category`, `ImageFile` are
removed from `Product`, `ProductConfigurations` and related
methods/handlers
- `Stock` and `Threshold` now hold integer values
- New initial migration was created

### Notes / Edge cases
This refactor was needed due to a lot of noise coming from the
unnecessary fields making it difficult to focus on MVP. Some of these
might be added as part full features down the line.

### Checklist
- [x] I kept the change focused (no unrelated refactors)
- [ ] Tests added/updated (or N/A with reason)
- [x] Docs updated (or N/A)
- [x] No secrets or environment-specific overrides committed

+++ apps/api/Modules/Inventory/Inventory/Products/Features/SetProductThreshold/SetProductThresholdHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace Inventory.Products.Features.SetProductThreshold;
 
-public record SetProductThresholdCommand(Guid Id, decimal Threshold)
+public record SetProductThresholdCommand(Guid Id, int Threshold)
     : ICommand<SetProductThresholdResult>;
 
 public record SetProductThresholdResult(Guid Id);
